fixed derivation of QuickCheck 1 and 2's Arbitrary Class required annotating the *raw* output of the 'guess' function, because guess does not understand type annotations e.g. 'choose (0::Int,3)' in the examples that you can create the guesses from.
this is necessary because otherwise the typechecker gets an ambiguous type variable when
dealing with 'choose' and it's basically not possible to derive Arbitrary instances for your types.

because of that this is sort of a hack; but it's a hack that makes automatic derivations
for Arbitrary actually work and that's good. neil might want to look into this later.